fix the word select
