+++ packages/api/libs/controller/crossword.ts
@@ -16,6 +16,7 @@ interface CrosswordResponse {
 	crossword: string[][];
 	title: string;
 	metadata: WordData[];
+	id: number;
 }
 
 async function getCrosswordDetails(
@@ -45,9 +46,10 @@ async function getRandomCrossword(
 
 		const words = randomCrossword.crosswordWords.map((v) => v.words.word_text);
 		const [crossword, metadata] = generateCrossword(words);
-
+	
 		const response: CrosswordResponse = {
 			title: randomCrossword?.title,
+			id: randomCrossword.crossword_id,
 			metadata: metadata.words_data.map((data) => {
 				const definition = randomCrossword.crosswordWords.find(
 					(word) => word.words.word_text === data.word,
