+++ packages/web-frontend/src/interfaces/crossword.tsx
@@ -1,7 +1,10 @@
+type Direction = "horizontal" | "vertical";
+
+
 export interface WordData {
 	word: string;
 	start_row: number;
 	start_col: number;
 	definition: string;
-	direction: "horizontal" | "vertical";
+	direction: Direction;
 }
